Change timelapse submission so that TimelapseProcess is in views.py
This unifies timelapse creation with pipeline process submission,
working towards #6. It should now be possible to catch redis/dramatiq
errors in views.py to set AsyncProcess status and failure message.

This commit also extends the timelapse tests, and improves the tests to
create media files in temporary directories on a per-test case basis.